chore: link app pages to matching modules
